# MobyPark/OpenAPI/openapi.yaml
openapi: 3.0.4
info: {}
paths:
...
  /api/Payments:
    post:
      tags:
      - Payments
      summary: Creates a pending payment and associated transaction.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreatePaymentDto'
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Internal processing error
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDto'
  /api/Payments/refund:
    post:
      tags:
      - Payments
      summary: Refunds an existing payment (Admin only).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRefundDto'
          text/json:
            schema:
              $ref: '#/components/schemas/PaymentRefundDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PaymentRefundDto'
      responses:
        '400':
          description: Invalid input or missing amount
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '201':
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponseDto'
        '404':
          description: Original payment not found
  /api/Payments/{paymentId}:
    put:
      tags:
      - Payments
      summary: Validates and completes a payment.
      parameters:
      - name: paymentId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDataDto'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionDataDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionDataDto'
      responses:
        '400':
          description: Invalid ID or data
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Payment validated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponseDto'
        '404':
          description: Payment not found
    delete:
      tags:
      - Payments
      summary: Deletes a payment record.
      parameters:
      - name: paymentId
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema: {}
        '404':
          description: Payment not found
  /api/Payments/{username}:
    get:
      tags:
      - Payments
      summary: Retrieves all payments for a specific user (Admin only).
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentModel'
        '404':
          description: User or payments not found
  /api/Payments/id/{paymentId}:
    get:
      tags:
      - Payments
      summary: Retrieves a payment by its unique ID.
      parameters:
      - name: paymentId
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid ID format
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentModel'
        '404':
          description: Payment not found
  /api/Payments/transaction/{transactionId}:
    get:
      tags:
      - Payments
      summary: Retrieves a payment by its associated transaction ID.
      parameters:
      - name: transactionId
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid ID format
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentModel'
        '404':
          description: Payment not found
  /api/Payments/plate/{licensePlate}:
    get:
      tags:
      - Payments
      summary: Retrieves all payments associated with a license plate.
      parameters:
      - name: licensePlate
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentModel'
        '404':
          description: No payments found
...
components:
  schemas:
...
    CreatePaymentDto:
      required:
      - amount
      - licensePlateNumber
      type: object
      properties:
        amount:
          type: number
          description: The monetary amount to be charged.
          format: double
        licensePlateNumber:
          minLength: 1
          type: string
          description: The license plate number associated with the payment.
      additionalProperties: false
      description: Data required to initiate a new payment.
...
    LicensePlateModel:
      required:
      - licensePlateNumber
      type: object
      properties:
        licensePlateNumber:
          minLength: 1
          type: string
      additionalProperties: false
...
    PaymentModel:
      required:
      - amount
      - createdAt
      - licensePlateNumber
      - transactionId
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        licensePlateNumber:
          minLength: 1
          type: string
        licensePlate:
          $ref: '#/components/schemas/LicensePlateModel'
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        transactionId:
          type: string
          format: uuid
        transaction:
          $ref: '#/components/schemas/TransactionModel'
      additionalProperties: false
    PaymentRefundDto:
      required:
      - paymentId
      type: object
      properties:
        paymentId:
          minLength: 1
          type: string
          description: The unique identifier of the payment to be refunded.
        amount:
          type: number
          description: The amount to refund. If null, may default to full amount.
          format: double
          nullable: true
      additionalProperties: false
      description: Data required to process a refund.
    PaymentResponseDto:
      type: object
      properties:
        status:
          type: string
          description: The outcome status of the operation (e.g., 'Success', 'Created').
          nullable: true
        payment:
          $ref: '#/components/schemas/PaymentModel'
      additionalProperties: false
      description: Wraps the payment model with a status message.
...
    RefundResponseDto:
      type: object
      properties:
        status:
          type: string
          description: The outcome status of the refund operation.
          nullable: true
        refund:
          $ref: '#/components/schemas/PaymentModel'
      additionalProperties: false
      description: Details regarding a processed refund.
...
    TransactionDataDto:
      required:
      - bank
      - issuer
      - method
      type: object
      properties:
        method:
          minLength: 1
          type: string
          description: The payment method used (e.g., CreditCard, PayPal).
        issuer:
          minLength: 1
          type: string
          description: The issuer of the payment method (e.g., Visa, Mastercard).
        bank:
          minLength: 1
          type: string
          description: The bank or provider handling the transaction.
      additionalProperties: false
      description: Details regarding the transaction method provided by the payment
        gateway.
    TransactionModel:
      required:
      - amount
      - bank
      - issuer
      - method
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        method:
          minLength: 1
          type: string
        issuer:
          minLength: 1
          type: string
        bank:
          minLength: 1
          type: string
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Please enter a valid token (e.g., ''Bearer eyJ...'').

        You can obtain a token by logging in or registering a new account via the
        endpoints.'
      scheme: Bearer
      bearerFormat: JWT
security:
- Bearer: []